+++ src/bin/bat/app.rs
@@ -405,6 +405,7 @@ impl App {
                 } else {
                     match self.matches.get_one::<String>("wrap").map(|s| s.as_str()) {
                         Some("character") => WrappingMode::Character,
+                        Some("word") => WrappingMode::Word,
                         Some("never") => WrappingMode::NoWrapping(true),
                         Some("auto") | None => {
                             if self.interactive_output || maybe_term_width.is_some() {